Move hooks into folder
